src/requirements.txt
asyncpg==0.27.0
databases[postgresql]==0.7.0
fastapi==0.89.1
psycopg2-binary==2.9.5
SQLAlchemy==1.4.46
uvicorn[standard]==0.20.0
python-dotenv==0.21.0
pydantic==1.10.4
Jinja2==3.1.2
requests==2.28.2
# dev
pytest==7.2.0
httpx==0.23.3